refactor(arch): use reexport instead of arch-specific paths
