+++ src/arch/aarch64/kernel/mod.rs
@@ -20,8 +20,8 @@ use core::sync::atomic::{AtomicPtr, AtomicU32, Ordering};
 
 pub(crate) use self::interrupts::wakeup_core;
 pub(crate) use self::processor::set_oneshot_timer;
-use crate::arch::aarch64::kernel::core_local::*;
-use crate::arch::aarch64::mm::paging::{BasePageSize, PageSize};
+use crate::arch::kernel::core_local::*;
+use crate::arch::mm::paging::{BasePageSize, PageSize};
 use crate::config::*;
 
 #[repr(align(8))]
@@ -110,7 +110,7 @@ pub fn boot_next_processor() {
 
 		use memory_addresses::VirtAddr;
 
-		use crate::arch::aarch64::start::smp::{TTBR0, smp_start};
+		use crate::arch::start::smp::{TTBR0, smp_start};
 		use crate::mm::virtual_to_physical;
 
 		if cpu_online == 0 {
