+++ src/arch/riscv64/start/hermit_entry.rs
@@ -4,10 +4,11 @@ use core::sync::atomic::Ordering;
 use hermit_entry::Entry;
 use hermit_entry::boot_info::RawBootInfo;
 
-use crate::arch::kernel::{CPU_ONLINE, CURRENT_BOOT_ID, HART_MASK, NUM_CPUS};
-use crate::arch::riscv64::kernel::CURRENT_STACK_ADDRESS;
 #[cfg(not(feature = "smp"))]
-use crate::arch::riscv64::kernel::processor;
+use crate::arch::kernel::processor;
+use crate::arch::kernel::{
+	CPU_ONLINE, CURRENT_BOOT_ID, CURRENT_STACK_ADDRESS, HART_MASK, NUM_CPUS,
+};
 use crate::config::KERNEL_STACK_SIZE;
 use crate::env;
