+++ src/arch/x86_64/kernel/mmio.rs
@@ -16,8 +16,8 @@ use memory_addresses::{PhysAddr, VirtAddr};
 use virtio::mmio::{DeviceRegisters, DeviceRegistersVolatileFieldAccess};
 use volatile::VolatileRef;
 
-use crate::arch::x86_64::mm::paging;
-use crate::arch::x86_64::mm::paging::{
+use crate::arch::mm::paging;
+use crate::arch::mm::paging::{
 	BasePageSize, PageSize, PageTableEntryFlags, PageTableEntryFlagsExt,
 };
 use crate::drivers::InterruptHandlerMap;
